Relocate entity package
