+++ app/domain/entity/user.py
@@ -3,8 +3,8 @@ from datetime import datetime
 from sqlalchemy import Column, String, DateTime, func
 from sqlalchemy.orm import Mapped, relationship
 
-from app.entity import Base
-from app.entity.util import generate_uuid
+from app.domain.entity import Base
+from app.domain.entity.util import generate_uuid
 
 
 class User(Base):
