+++ app/domain/entity/util.py
