+++ pyproject.toml
@@ -19,6 +19,7 @@ pyjwt = "^2.8.0"
 pytest = "^7.4.3"
 pytest-cov = "^4.1.0"
 redis = "^5.0.1"
+httpx = "^0.26.0"
 
 
 [build-system]
